Progress on Major Refactor
Goal is to have a standard way to render things. The strategy is to have
a Mesh struct which holds all of the vertex buffers, and an Entity which
holds information about the game world. The two of them will be combined
in a Model struct, which will use them to render. This Mesh-Entity split
should make it easy to deal with the game world and rendering
separately. Currently it's mostly working. I can draw textures, colors,
normals, and normal vertices all via the same shader - and more
importantly, if some of them aren't specified, there is a default
behavior so the model will still be rendered. I'm working on
reimplementing the shapes that were previously in geom/ and am most of
the way there.