Remove old shaders, make line a mesh
Put the axis class in utils and adjusted it to use the new line mesh.
I'd like to find a better place for it eventually.
Not too pleased with the axes now being somewhat dark - due to using the
ambient lighting in the scene rather than just a fixed color.
Also added an entity.Default and a log message when trying to draw an
entity with scale{0,0,0} since I kept forgetting to set scale, which
made nothing appear.