Use Quaternions for Rotation
Previously used a vec3 for roll, pitch, and yaw. Quaternions were always
a TODO so finally got to it. I'm still not entirely satisfied since
quaternions are confusing and I feel as though they are now too visible
to developers. I'd like to add more methods so people can avoid working
with them if they don't want to.