Allow modification of Center of loaded OBJ meshes
Ideally meshes are created with their intended center point at the
origin, but that may not be the case. This addition allows users to
specify where the center is as a percent of the rectangular prism
bounding box that hold the mesh. It's a bit clunky. There's not really a
great way for a user to find what percentages they should use without
some code, or using a modeling program. Ideally any models should
already have their center at the origin so it won't be an issue and no
one will need this...