Improve & Restructure Camera package
Made camera a base struct which the others embed. This removes a lot of
duplicate code and makes it much easier to add new cameras. Added a
FreeCamera which has 5 degrees of movement(X,Y,Z motion and Pitch+Yaw)
hardcoded in.  I plan on making a better input controller class so it
doesn't depend on specific keybinds. Also it seems to work, but I think
the math is incorrect for determining the Up vector for the
TargetCamera.